♻️ [timeago/date-countdown/date-display] Split component files for packaging (#32440)
* Split amp-timeago component files

* Split amp-date-countdown component files

* Split amp-date-display component files

* Move updatePropsForRendering into BaseElement

* Allowlist timeago component.js for date math

* Rename allowlisted timeago dependencies on third_party/timeagojs
