+++ build-system/tasks/presubmit-checks.js
@@ -1061,7 +1061,7 @@ const forbiddenTermsSrcInclusive = {
     message: 'Unless you do weird date math (allowlist), use Date.now().',
     allowlist: [
       'extensions/amp-timeago/0.1/amp-timeago.js',
-      'extensions/amp-timeago/1.0/timeago.js',
+      'extensions/amp-timeago/1.0/component.js',
       'src/utils/date.js',
     ],
   },
