+++ extensions/amp-date-countdown/1.0/component.type.js
