+++ extensions/amp-timeago/1.0/amp-timeago.js
@@ -14,16 +14,14 @@
  * limitations under the License.
  */
 
-import {PreactBaseElement} from '../../../src/preact/base-element';
-import {Timeago} from './timeago';
+import {BaseElement} from './base-element';
 import {isExperimentOn} from '../../../src/experiments';
-import {parseDateAttrs as parseDateAttrsBase} from '../../../src/utils/date';
 import {userAssert} from '../../../src/log';
 
 /** @const {string} */
 const TAG = 'amp-timeago';
 
-class AmpTimeago extends PreactBaseElement {
+class AmpTimeago extends BaseElement {
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
@@ -33,44 +31,6 @@ class AmpTimeago extends PreactBaseElement {
     );
     return super.isLayoutSupported(layout);
   }
-
-  /** @override */
-  updatePropsForRendering(props) {
-    props['placeholder'] = props['children'];
-  }
-}
-
-/** @override */
-AmpTimeago['Component'] = Timeago;
-
-/** @override */
-AmpTimeago['passthroughNonEmpty'] = true;
-
-/** @override */
-AmpTimeago['layoutSizeDefined'] = true;
-
-/** @override */
-AmpTimeago['props'] = {
-  'datetime': {
-    attrs: ['datetime', 'timestamp-ms', 'timestamp-seconds', 'offset-seconds'],
-    parseAttrs: parseDateAttrs,
-  },
-  'cutoff': {attr: 'cutoff', type: 'number'},
-  'locale': {attr: 'locale'},
-};
-
-/**
- * @param {!Element} element
- * @return {?number}
- * @throws {UserError} when attribute values are missing or invalid.
- * @visibleForTesting
- */
-export function parseDateAttrs(element) {
-  return parseDateAttrsBase(element, [
-    'datetime',
-    'timestamp-ms',
-    'timestamp-seconds',
-  ]);
 }
 
 AMP.extension(TAG, '1.0', (AMP) => {
