+++ extensions/amp-timeago/1.0/component.js
